Rename all the modules/templates that make up `<hypothesis-app>`
Rename the modules and templates for the `<hypothesis-app>` component to
match the component name, following our conventions.
