--- src/sidebar/directive/app.js
@@ -1,9 +1,0 @@
-'use strict';
-
-var AppController = require('../app-controller');
-
-module.exports = {
-  controllerAs: 'vm',
-  controller: AppController,
-  template: require('../templates/app.html'),
-};
