+++ src/sidebar/templates/hypothesis_app.html
